+++ fvm/src/lib.rs
@@ -55,6 +55,7 @@ mod test {
     use crate::call_manager::DefaultCallManager;
     use crate::engine::EnginePool;
     use crate::externs::{Chain, Consensus, Externs, Rand};
+    use crate::kernel::filecoin::DefaultFilecoinKernel;
     use crate::machine::{DefaultMachine, Manifest, NetworkConfig};
     use crate::state_tree::StateTree;
     use crate::{executor, DefaultKernel};
@@ -124,9 +125,8 @@ mod test {
 
         let machine = DefaultMachine::new(&mc, bs, DummyExterns).unwrap();
         let engine = EnginePool::new_default((&mc.network).into()).unwrap();
-        let _ = executor::DefaultExecutor::<DefaultKernel<DefaultCallManager<_>>>::new(
-            engine,
-            Box::new(machine),
-        );
+        let _ = executor::DefaultExecutor::<
+            DefaultFilecoinKernel<DefaultKernel<DefaultCallManager<_>>>,
+        >::new(engine, Box::new(machine));
     }
 }
