+++ testing/test_actors/actors/fil-syscall-actor/src/actor.rs
@@ -40,7 +40,6 @@ pub fn invoke(_: u32) -> u32 {
     test_message_context();
     test_balance();
     test_unaligned();
-    test_upgrade();
 
     #[cfg(coverage)]
     sdk::debug::store_artifact("syscall_actor.profraw", minicov::capture_coverage());
@@ -377,16 +376,4 @@ fn test_unaligned() {
         let expected = context().unwrap();
         assert_eq!(expected, actual);
     }
-}
-
-fn test_upgrade() {
-    // test that calling `upgrade_actor` on ourselves results in a SYS_INVALID_RECEIVER error
-    // since we don't have a upgrade endpoint
-    let code_cid = sdk::actor::get_actor_code_cid(&Address::new_id(10000)).unwrap();
-    let res = sdk::actor::upgrade_actor(&code_cid, None).unwrap();
-
-    assert_eq!(
-        res.exit_code,
-        fvm_shared::error::ExitCode::SYS_INVALID_RECEIVER,
-    );
 }
